requirements.txt
streamlit
google_generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
sentence-transformers
langchain-community
langchain-huggingface
